Agent processor and schema refactor WIP
