+++ ciris_engine/services/llm_client.py
@@ -10,7 +10,7 @@ import instructor
 
 # Configuration
 from ciris_engine.core.config_manager import get_config
-from ciris_engine.core.config_schemas import OpenAIConfig
+from ciris_engine.schemas.config_schemas_v1 import OpenAIConfig # ERIC
 
 logger = logging.getLogger(__name__)
