+++ ciris_engine/services/llm_service.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 from .base import Service
 from ciris_engine.services.llm_client import CIRISLLMClient
-from ciris_engine.core.config_schemas import OpenAIConfig, LLMServicesConfig
+from ciris_engine.schemas.config_schemas_v1 import OpenAIConfig, LLMServicesConfig # ERIC
 
 logger = logging.getLogger(__name__)
